libc: Update include arch to last version
This code is being updated out of the tree of scc and
it is time to synchroniza both copies now.

+++ include/bits/ppc/arch/cdefs.h
@@ -1,6 +1,6 @@
 #ifdef _NEED_SIZET
 #ifndef _SIZET
-typedef unsigned long size_t;
+typedef unsigned size_t;
 #define _SIZET
 #endif
 #undef _NEED_SIZET
